v2.1 was a dud. Score 0.52870
and major step backward in accuracy. Reverted to the Random Forest model in v2.0 and found that `Survived` in the test set needs to be encoded as a factor before training the model, from its current state as numeric. The consequence of this type mismatch is the model is attempting to use regression on a continuous numeric variable with a range of 0 to 1. In reality, this is a classification problem. Since v2.0 scored .76315 using rounded results from regression instead of classification, fixing this problem should result in a large increase in model accuracy relative to the 76% level in v2.0.